Added session joining functionality

+++ src/group-session.js
@@ -19,9 +19,15 @@
     let isEnabled;
     // Option: Show scan code in full screen mode
     let showScanCode;
+    // Whether user is about to join a session (by entering link)
+    let isJoiningSession;
+    // Whether user just triggered a link copy
+    let isCopyingLink;
     // Private session id
     let session_id;
-    // public session token used for others to join
+    // Whether user is the owner of the group session
+    let is_session_owner;
+    // Public session token used for others to join
     let join_session_token;
     // Contains the interval id
     let alive_job;
@@ -44,15 +50,58 @@
         // Doesn't fail. If session already exists, the current session is returned.
         try {
             const res_join = await Spicetify.CosmosAsync.get(`https://spclient.wg.spotify.com/social-connect/v2/sessions/current_or_new?local_device_id=${local_device_id}&type=REMOTE`);
-            
             session_id = res_join["session_id"];
             join_session_token = res_join["join_session_token"];
+            is_session_owner = res_join["is_session_owner"];
             users = res_join["session_members"];
             alive_job = setInterval(checkAlive, ALIVE_INTERVAL);
             imageCache = {};
         } catch (e) {
             Spicetify.showNotification("Session creation failed. Make sure your connected to the internet and the account has Spotify Premium.");
         }
+    }
+
+    const joinSession = async (sessionId) => {
+        const local_device_id = Spicetify.Player.data.play_origin.device_identifier;
+        if (local_device_id == null) {
+            Spicetify.showNotification("Local device id is unknown. Try to play music before creating a new session.");
+            return;
+        }
+
+        // Doesn't fail. If session already exists, the current session is returned.
+        try {
+            const res_join = await Spicetify.CosmosAsync.post(`https://spclient.wg.spotify.com/social-connect/v2/sessions/join/${sessionId}/?join_type=deeplinking&local_device_id=${local_device_id}&playback_control=listen_and_control`);
+            session_id = res_join["session_id"];
+            join_session_token = res_join["join_session_token"];
+            is_session_owner = res_join["is_session_owner"];
+            users = res_join["session_members"];
+            alive_job = setInterval(checkAlive, ALIVE_INTERVAL);
+            imageCache = {};
+        } catch (e) {
+            Spicetify.showNotification("Session joining failed. Make sure your connected to the internet and the account has Spotify Premium.");
+        }
+    }
+
+    const leaveSession = async () => {
+        const local_device_id = Spicetify.Player.data.play_origin.device_identifier;
+
+        if (local_device_id == null) {
+            Spicetify.showNotification("Local device id is unknown.");
+            return;
+        }
+
+        if (session_id != null) {
+            // On success, the response is empty.
+            // On error, it contains error_type and message.
+            const res_leave = await Spicetify.CosmosAsync.post(`https://spclient.wg.spotify.com/social-connect/v3/sessions/${session_id}/leave?local_device_id=${local_device_id}`);
+
+            if ("error_type" in res_leave) {
+                Spicetify.showNotification(res_leave.message);
+            }
+        }
+        
+        // cleanup
+        clearGlobals();
     }
 
     const deleteSession = async () => {
@@ -195,6 +244,8 @@
         titleDiv.appendChild(title);
         title.classList.add("connect-title__text");
         title.textContent = "Group Session";
+        title.style.padding = "0px 0px";
+        title.style.marginBottom = "3px";
 
         const helpLink = document.createElement("a");
         titleDiv.appendChild(helpLink);
@@ -213,28 +264,77 @@
         return containerDiv;
     }
 
-    const buttonStart = async (e) => {
+    const timeout = (ms) => {
+        return new Promise(resolve => setTimeout(resolve, ms));
+    }
+
+    const handleCreateButtonPressed = async (e) => {
         e.target.disabled = true;
         await createSession();
+        updateMenu();
+    }
+
+    const handleJoinButtonPressed = async (e) => {
+        e.target.disabled = true;
+        isJoiningSession = true;
+        await timeout(100);
         updateMenu();
     }
 
     const buildStartMenu = () => {
         const containerDiv = buildContainerAndTitle();
         
+        const buttonDiv = document.createElement("div");
+        buttonDiv.style.display = "flex";
+        buttonDiv.style.marginTop = "5px";
+
         const createButton = document.createElement("button");
+        createButton.style.flex = "1";
+        createButton.style.marginRight = "7px";
+        createButton.style.padding = "8px 8px";
         createButton.classList.add("main-buttons-button", "main-button-primary");
         createButton.textContent = "Start Session";
-        createButton.addEventListener("click", buttonStart);
-        containerDiv.appendChild(createButton);
+        createButton.addEventListener("click", handleCreateButtonPressed);
+        buttonDiv.appendChild(createButton);
+
+        const joinButton = document.createElement("button");
+        joinButton.style.flex = "1";
+        joinButton.style.padding = "8px 8px";
+        joinButton.classList.add("main-buttons-button", "main-button-primary");
+        joinButton.textContent = "Join Session";
+        joinButton.addEventListener("click", handleJoinButtonPressed);
+        buttonDiv.appendChild(joinButton);
+
+        containerDiv.appendChild(buttonDiv);
 
         return containerDiv;
     }
 
-    const buttonLeave = async (e) => {
+    const handleLeaveButtonPressed = async (e) => {
         e.target.disabled = true;
-        await deleteSession();
+        if (is_session_owner) {
+            await deleteSession();
+        } else {
+            await leaveSession();
+        }
         updateMenu();
+    }
+
+    const handleCopyLink = (e) => {
+        if (!isCopyingLink) {
+            return;
+        }
+        // @ts-ignore
+        const sessionLink = document.getElementById("spicetify-group-session-menu-link").value;
+        e.clipboardData.setData('text/plain', sessionLink);
+        e.preventDefault();
+        Spicetify.showNotification("Link copied to clipboard");
+        isCopyingLink = false;
+    }
+
+    const handleCopyLinkPressed = () => {
+        isCopyingLink = true;
+        document.execCommand('copy');
     }
 
     const buildSessionMenu = () => {
@@ -255,9 +355,11 @@
         copyLink.id = "spicetify-group-session-menu-link";
         copyLink.type = "text";
         copyLink.readOnly = true;
+        copyLink.onclick = handleCopyLinkPressed;
         copyLink.value = `https://open.spotify.com/socialsession/${join_session_token}`;
         copyLink.classList.add("main-playlistEditDetailsModal-textElement", "main-playlistEditDetailsModal-titleInput");
         copyLink.style.marginBottom = "12px";
+        copyLink.style.cursor = "pointer";
         containerDiv.appendChild(copyLink);
 
         const listenerButton = document.createElement("button");
@@ -268,7 +370,7 @@
         containerDiv.appendChild(listenerButton);
 
         const closeButton = document.createElement("button");
-        closeButton.addEventListener("click", buttonLeave);
+        closeButton.addEventListener("click", handleLeaveButtonPressed);
         closeButton.classList.add("main-buttons-button", "main-button-outlined");
         closeButton.textContent = "Close Session";
         closeButton.style.fontSize = "8px";
@@ -278,11 +380,85 @@
         return containerDiv;
     }
 
+    const extractSessionId = (sessionLink) => {
+        const prefix = "open.spotify.com/socialsession/";
+        let startIndex = sessionLink.indexOf(prefix);
+        if (startIndex === -1) { // Obviously invalid link
+            return null;
+        }
+        startIndex += prefix.length;
+        let endIndex = sessionLink.lastIndexOf("?");
+        if (endIndex === -1) {
+            endIndex = sessionLink.length;
+        }
+        return sessionLink.substring(startIndex, endIndex);
+    }
+
+    const handleConfirmJoinButtonPressed = async (e) => {
+        e.target.disabled = true;
+
+        // Reading link
+        const enterLinkField = document.getElementById("spicetify-group-session-join-link");
+        
+        // @ts-ignore
+        const sessionId = extractSessionId(enterLinkField.value);
+        if (sessionId !== null) {
+            await joinSession(sessionId);
+            isJoiningSession = false;
+        } else {
+            Spicetify.showNotification("Could not join session. Maybe the session link is invalid?");
+            await timeout(100);
+        }
+        updateMenu();
+    }
+
+    const handleCancelJoinButtonPressed = async (e) => {
+        e.target.disabled = true;
+        isJoiningSession = false;
+        await timeout(100);
+        updateMenu();
+    }
+
+    const buildJoinMenu = () => {
+        const containerDiv = buildContainerAndTitle();
+
+        const enterLink = document.createElement("input");
+        enterLink.id = "spicetify-group-session-join-link";
+        enterLink.type = "text";
+        enterLink.onkeyup = (e) => {
+            if (e.key === "Enter") {
+                handleConfirmJoinButtonPressed();
+            }
+        };
+        enterLink.placeholder = `https://open.spotify.com/socialsession/`;
+        enterLink.classList.add("main-playlistEditDetailsModal-textElement", "main-playlistEditDetailsModal-titleInput");
+        enterLink.style.marginBottom = "12px";
+        setTimeout(() => enterLink.focus(), 100);
+        containerDiv.appendChild(enterLink);
+
+        const confirmJoinButton = document.createElement("button");
+        confirmJoinButton.addEventListener("click", handleConfirmJoinButtonPressed);
+        confirmJoinButton.classList.add("main-buttons-button", "main-button-outlined");
+        confirmJoinButton.textContent = "Join";
+        confirmJoinButton.style.marginBottom = "12px";
+        confirmJoinButton.style.marginRight = "5px";
+        containerDiv.appendChild(confirmJoinButton);
+
+        const cancelButton = document.createElement("button");
+        cancelButton.addEventListener("click", handleCancelJoinButtonPressed);
+        cancelButton.classList.add("main-buttons-button", "main-button-outlined");
+        cancelButton.textContent = "Cancel";
+        cancelButton.style.marginBottom = "12px";
+        containerDiv.appendChild(cancelButton);
+        
+        return containerDiv;
+    };
+
     const updateMenu = () => {
         // remove the old menu if exists
-        const old_menu = document.getElementById("spicetify-group-session-menu");
-        if (old_menu != null) {
-            old_menu.remove();
+        const oldMenu = document.getElementById("spicetify-group-session-menu");
+        if (oldMenu != null) {
+            oldMenu.remove();
         }
 
         const deviceMenu = document.querySelector(".connect-device-list-content");
@@ -292,10 +468,13 @@
 
         // get the new menu
         let containerDiv;
-        if (session_id != null && join_session_token != null) {
+        if (session_id != null && join_session_token != null && is_session_owner != null) {
             containerDiv = buildSessionMenu();
+        } else if (isJoiningSession) {
+            containerDiv = buildJoinMenu();
         } else {
             session_id = null;
+            is_session_owner = null;
             join_session_token = null;
             containerDiv = buildStartMenu();
         }
@@ -350,13 +529,17 @@
 
         const target = e.target;
         // Make sure we accutally have an element
-        if (target == null) {
+        if (target == null || !("classList" in target)) {
             return;
         }
 
-        // Update menu on display of the device menu
+        // Update menu on display of the device menu, and when devices appear & disappear
         // @ts-ignore
-        if ("classList" in target && target.classList.contains("connect-device-list-container--is-visible")) {
+        if (target.classList.contains("connect-device-list-container--is-visible") || 
+            target.classList.contains("connect-device-list") ||
+            ("previousSibling" in target && target.previousSibling !== null && 
+             target.previousSibling.id === "spicetify-group-session-menu")
+        ) {
             updateMenu();
             return;
         }
@@ -364,6 +547,7 @@
 
     const clearGlobals = () => {
         session_id = null;
+        is_session_owner = null;
         join_session_token = null;
         imageCache = null;
         users = null;
@@ -373,12 +557,14 @@
     const initialize = () => {
         injectListenerStyles();
         document.addEventListener("DOMNodeInserted", insertObserver);
+        document.addEventListener("copy", handleCopyLink);
         const fs_btn = document.querySelector(".main-nowPlayingBar-right > .ExtraControls > button.control-button");
         fs_btn.addEventListener("click", onFullscreen);
     }
 
     const terminate = () => {
         document.removeEventListener("DOMNodeInserted", insertObserver);
+        document.removeEventListener("copy", handleCopyLink);
         const fs_btn = document.querySelector(".main-nowPlayingBar-right > .ExtraControls > button.control-button");
         fs_btn.addEventListener("click", onFullscreen);
         clearGlobals();
@@ -387,6 +573,8 @@
 
     isEnabled = Spicetify.LocalStorage.get("group-session-enabled") === "true" || Spicetify.LocalStorage.get("group-session-enabled") === null;
     showScanCode = Spicetify.LocalStorage.get("group-session-showscancode") === "true";
+    isJoiningSession = false;
+    isCopyingLink = false;
 
     if (isEnabled) {
         initialize();
